Squeeze tests
